+++ packages/models/src/Domain/Syncable/SmartView/SmartViewBuilder.ts
@@ -74,10 +74,22 @@ export function BuildSmartViews(
     }),
   )
 
+  const starred = new SmartView(
+    new DecryptedPayload({
+      uuid: SystemViewId.StarredNotes,
+      content_type: ContentType.SmartView,
+      ...PayloadTimestampDefaults(),
+      content: FillItemContent<SmartViewContent>({
+        title: 'Starred',
+        predicate: starredNotesPredicate(options).toJson(),
+      }),
+    }),
+  )
+
   if (supportsFileNavigation) {
-    return [notes, files, archived, trash, untagged]
+    return [notes, starred, files, archived, trash, untagged]
   } else {
-    return [notes, archived, trash, untagged]
+    return [notes, starred, archived, trash, untagged]
   }
 }
 
@@ -166,6 +178,25 @@ function untaggedNotesPredicate(options: FilterDisplayOptions) {
   ]
   if (options.includeArchived === false) {
     subPredicates.push(new Predicate('archived', '=', false))
+  }
+  if (options.includeProtected === false) {
+    subPredicates.push(new Predicate('protected', '=', false))
+  }
+  if (options.includePinned === false) {
+    subPredicates.push(new Predicate('pinned', '=', false))
+  }
+  const predicate = new CompoundPredicate('and', subPredicates)
+
+  return predicate
+}
+
+function starredNotesPredicate(options: FilterDisplayOptions) {
+  const subPredicates: Predicate<SNNote>[] = [
+    new Predicate('starred', '=', true),
+    new Predicate('content_type', '=', ContentType.Note),
+  ]
+  if (options.includeTrashed === false) {
+    subPredicates.push(new Predicate('trashed', '=', false))
   }
   if (options.includeProtected === false) {
     subPredicates.push(new Predicate('protected', '=', false))