+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.148](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.147...@standardnotes/releases@1.3.148) (2022-10-17)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.147](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.146...@standardnotes/releases@1.3.147) (2022-10-14)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.3.146](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.145...@standardnotes/releases@1.3.146) (2022-10-14)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.3.145](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.3.144...@standardnotes/releases@1.3.145) (2022-10-13)
 
 **Note:** Version bump only for package @standardnotes/releases